+++ Socolin.RabbitMQ.Client/ConsumerPipes/Builders/DefaultConsumerPipeBuilder.cs
@@ -1,0 +1,19 @@
+using System;
+
+namespace Socolin.RabbitMQ.Client.ConsumerPipes.Builders
+{
+	public class DefaultConsumerPipeBuilder<T> : IConsumerPipeBuilder<T> where T : class
+	{
+		private readonly Func<IConsumerPipe<T>> _build;
+
+		public DefaultConsumerPipeBuilder(Func<IConsumerPipe<T>> build)
+		{
+			_build = build;
+		}
+
+		public IConsumerPipe<T> Build()
+		{
+			return _build.Invoke();
+		}
+	}
+}