+++ Socolin.RabbitMQ.Client/ConsumerPipes/Builders/IConsumerPipeBuilder.cs
@@ -1,0 +1,7 @@
+namespace Socolin.RabbitMQ.Client.ConsumerPipes.Builders
+{
+	public interface IConsumerPipeBuilder<T> where T : class
+	{
+		IConsumerPipe<T> Build();
+	}
+}