+++ Socolin.RabbitMQ.Client/Options/RabbitMqServiceOptionsBuilder.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using System.Collections.ObjectModel;
 using Socolin.RabbitMQ.Client.Exceptions;
 using Socolin.RabbitMQ.Client.Pipes.Builders;
 
@@ -17,9 +16,6 @@ namespace Socolin.RabbitMQ.Client.Options
 
 		private Func<object, byte[]>? _serializer;
 		private string? _contentType;
-
-		private readonly IDictionary<string, Func<Type, ReadOnlyMemory<byte>, object>> _deserializers = new Dictionary<string, Func<Type, ReadOnlyMemory<byte>, object>>();
-		private Func<Type, ReadOnlyMemory<byte>, object>? _defaultDeserializer;
 
 		private readonly IList<IPipeBuilder> _customPipeBuilders = new List<IPipeBuilder>();
 
@@ -54,18 +50,6 @@ namespace Socolin.RabbitMQ.Client.Options
 			return this;
 		}
 
-		public RabbitMqServiceOptionsBuilder WithDefaultDeSerializer(Func<Type, ReadOnlyMemory<byte>, object> deserializer)
-		{
-			_defaultDeserializer = deserializer;
-			return this;
-		}
-
-		public RabbitMqServiceOptionsBuilder WithDeSerializer(Func<Type, ReadOnlyMemory<byte>, object> deserializer, string contentType)
-		{
-			_deserializers[contentType] = deserializer;
-			return this;
-		}
-
 		public RabbitMqServiceOptionsBuilder WithCustomPipe(IActionPipeBuilder pipeBuilder)
 		{
 			_customPipeBuilders.Add(pipeBuilder);
@@ -95,8 +79,6 @@ namespace Socolin.RabbitMQ.Client.Options
 
 			if (_serializer != null)
 				options.Serialization = new SerializationOption(_serializer, _contentType);
-			if (_defaultDeserializer != null)
-				options.Deserialization = new DeserializationGenericPipeOption(_defaultDeserializer, new ReadOnlyDictionary<string, Func<Type, ReadOnlyMemory<byte>, object>>(_deserializers));
 
 			if (_retryPipeBuilder != null)
 				options.Retry = _retryPipeBuilder;