+++ Socolin.RabbitMQ.Client/Pipes/Context/PipeContextMessage.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using RabbitMQ.Client;
 
 namespace Socolin.RabbitMQ.Client.Pipes.Context
@@ -9,6 +10,7 @@ namespace Socolin.RabbitMQ.Client.Pipes.Context
 		public string? QueueName { get; set; }
 		public IBasicProperties? BasicProperties { get; set; }
 		public ReadOnlyMemory<byte> SerializedMessage { get; set; }
+		public Dictionary<string, object> Items { get; } = new Dictionary<string, object>();
 
 		public PipeContextMessage(object message)
 		{